requirements.txt
matplotlib>=3.3,<3.4
psutil>=5.8,<5.9
tqdm>=4.59,<4.60
pandas>=1.1,<1.2
scipy>=1.5,<1.6
numpy>=1.10,<1.20
tensorboard==2.4.0
future==0.17.1
kornia
hydra-core
wget
seaborn
easydict
torchpq
git+https://github.com/lucasb-eyer/pydensecrf.git 
setuptools==59.5.0
pyDeprecate==0.3.1
torch==1.7.1
torchvision==0.8.2
scikit-image
opencv-python
pytorch-lightning
protobuf==3.20
python-dotenv
wandb
plotly
umap-learn
squirrel-core
hnne